refactor(onboard): move import state out of setup
